requirements.txt
web3
python-dotenv
pydantic
